+++ sentry-rails/lib/sentry/rails/active_job.rb
@@ -20,12 +20,12 @@ module Sentry
           def record(job, &block)
             Sentry.with_scope do |scope|
               begin
-                scope.set_transaction_name(job.class.name)
+                scope.set_transaction_name(job.class.name, source: :task)
                 transaction =
                   if job.is_a?(::Sentry::SendEventJob)
                     nil
                   else
-                    Sentry.start_transaction(name: scope.transaction_name, op: "active_job")
+                    Sentry.start_transaction(name: scope.transaction_name, source: scope.transaction_source, op: "active_job")
                   end
 
                 scope.set_span(transaction) if transaction