+++ sentry-ruby/lib/sentry/span.rb
@@ -104,6 +104,13 @@ module Sentry
       "#{@trace_id}-#{@span_id}-#{sampled_flag}"
     end
 
+    # Generates a W3C Baggage header string for distributed tracing
+    # from the incoming baggage stored on the transation.
+    # @return [String, nil]
+    def to_baggage
+      transaction&.get_baggage&.serialize
+    end
+
     # @return [Hash]
     def to_hash
       {